+++ providers/dns/liquidweb/liquidweb_integration_test.go
@@ -29,123 +29,8 @@ func setupTest(t *testing.T) (*DNSProvider, *http.ServeMux) {
 	return provider, mux
 }
 
-// func TestDNSProvider_Present(t *testing.T) {
-// 	provider, mux := setupTest(t)
-//
-// 	mux.HandleFunc("/v1/Network/DNS/Record/create", func(w http.ResponseWriter, r *http.Request) {
-// 		assert.Equal(t, http.MethodPost, r.Method)
-//
-// 		username, password, ok := r.BasicAuth()
-// 		assert.Equal(t, "blars", username)
-// 		assert.Equal(t, "tacoman", password)
-// 		assert.True(t, ok)
-//
-// 		reqBody, err := io.ReadAll(r.Body)
-// 		if err != nil {
-// 			http.Error(w, err.Error(), http.StatusInternalServerError)
-// 			return
-// 		}
-//
-// 		expectedReqBody := `
-// 			{
-// 				"params": {
-// 					"name": "_acme-challenge.tacoman.com",
-// 					"rdata": "\"47DEQpj8HBSa-_TImW-5JCeuQeRkm5NMpJWZG3hSuFU\"",
-// 					"ttl": 300,
-// 					"type": "TXT",
-// 					"zone": "tacoman.com"
-// 				}
-// 			}`
-// 		assert.JSONEq(t, expectedReqBody, string(reqBody))
-//
-// 		w.WriteHeader(http.StatusOK)
-// 		_, err = fmt.Fprintf(w, `{
-// 			"type": "TXT",
-// 			"name": "_acme-challenge.tacoman.com",
-// 			"rdata": "\"47DEQpj8HBSa-_TImW-5JCeuQeRkm5NMpJWZG3hSuFU\"",
-// 			"ttl": 300,
-// 			"id": 1234567,
-// 			"prio": null
-// 		}`)
-// 		if err != nil {
-// 			http.Error(w, err.Error(), http.StatusInternalServerError)
-// 			return
-// 		}
-// 	})
-//
-// 	err := provider.Present("tacoman.com", "", "")
-// 	require.NoError(t, err)
-// }
-//
-// func TestDNSProvider_CleanUp(t *testing.T) {
-// 	provider, mux := setupTest(t)
-//
-// 	mux.HandleFunc("/v1/Network/DNS/Record/delete", func(w http.ResponseWriter, r *http.Request) {
-// 		assert.Equal(t, http.MethodPost, r.Method)
-//
-// 		username, password, ok := r.BasicAuth()
-// 		assert.Equal(t, "blars", username)
-// 		assert.Equal(t, "tacoman", password)
-// 		assert.True(t, ok)
-//
-// 		_, err := fmt.Fprintf(w, `{"deleted": "123"}`)
-// 		if err != nil {
-// 			http.Error(w, err.Error(), http.StatusInternalServerError)
-// 			return
-// 		}
-// 	})
-//
-// 	provider.recordIDs["123"] = 1234567
-//
-// 	err := provider.CleanUp("tacoman.com.", "123", "")
-// 	require.NoError(t, err, "fail to remove TXT record")
-// }
-
-func TestDNSProvider_Present(t *testing.T) {
-	envTest.Apply(map[string]string{
-		EnvUsername: "blars",
-		EnvPassword: "tacoman",
-		EnvURL:      mockAPIServer(t),
-		EnvZone:     "tacoman.com", // this needs to be removed from test?
-	})
-
-	defer envTest.ClearEnv()
-
-	provider, err := NewDNSProvider()
-	require.NoError(t, err)
-
-	err = provider.Present("tacoman.com", "", "")
-	require.NoError(t, err)
-}
-
-func TestDNSProvider_CleanUp(t *testing.T) {
-	envTest.Apply(map[string]string{
-		EnvUsername: "blars",
-		EnvPassword: "tacoman",
-		EnvURL: mockAPIServer(t, network.DNSRecord{
-			Name:   "_acme-challenge.tacoman.com",
-			RData:  "123d==",
-			Type:   "TXT",
-			TTL:    300,
-			ID:     1234567,
-			ZoneID: 42,
-		}),
-		EnvZone: "tacoman.com", // this needs to be removed from test?
-	})
-
-	defer envTest.ClearEnv()
-
-	provider, err := NewDNSProvider()
-	require.NoError(t, err)
-
-	provider.recordIDs["123d=="] = 1234567
-
-	err = provider.CleanUp("tacoman.com.", "123d==", "")
-	require.NoError(t, err, "fail to remove TXT record")
-}
-
 // FIXME
-func TestIntegration(t *testing.T) {
+func TestFoo(t *testing.T) {
 	testCases := []struct {
 		desc          string
 		envVars       map[string]string
